Rename Instance to Cluster and move classes that are in Rubber::Configuration into lib/rubber/configuration to match namespace
