+++ lib/rubber/cloud/aws/base.rb
@@ -3,11 +3,11 @@ require 'rubber/cloud/aws'
 
 module Rubber
   module Cloud
-  
-    class Aws::Base < Fog
-      
+
+    class Aws::Base < ::Rubber::Cloud::Fog
+
       def initialize(env, capistrano)
-        
+
         compute_credentials = {
           :aws_access_key_id => env.access_key,
           :aws_secret_access_key => env.secret_access_key
@@ -21,7 +21,7 @@ module Rubber
         }
 
         @table_store = ::Fog::AWS::SimpleDB.new(compute_credentials)
-        
+
         compute_credentials[:region] = env.region
         @elb = ::Fog::AWS::ELB.new(compute_credentials)
 
@@ -33,9 +33,9 @@ module Rubber
         env['storage_credentials'] = storage_credentials
         super(env, capistrano)
       end
-      
+
       def table_store(table_key)
-        return Rubber::Cloud::Aws::TableStore.new(@table_store, table_key)  
+        return Rubber::Cloud::Aws::TableStore.new(@table_store, table_key)
       end
 
       def describe_instances(instance_id=nil)
@@ -116,7 +116,7 @@ module Rubber
           raise "Set #{k} in rubber.yml" unless "#{env[k]}".strip.size > 0
         end
         raise "create_image can only be called from a capistrano scope" unless capistrano
- 
+
         ec2_key = env.key_file
         ec2_pk = env.pk_file
         ec2_cert = env.cert_file
@@ -126,7 +126,7 @@ module Rubber
         ec2_cert_dest = "/mnt/#{File.basename(ec2_cert)}"
 
         storage(env.image_bucket).ensure_bucket
-        
+
         capistrano.put(File.read(ec2_key), ec2_key_dest)
         capistrano.put(File.read(ec2_pk), ec2_pk_dest)
         capistrano.put(File.read(ec2_cert), ec2_cert_dest)
@@ -146,7 +146,7 @@ module Rubber
             echo -n .
             sleep 5
           done
-          
+
           # this returns exit code even if pid has already died, and thus triggers fail fast shell error
           wait $bg_pid
         CMD
@@ -315,4 +315,3 @@ module Rubber
     end
   end
 end
-
