+++ lib/rubber/configuration.rb
@@ -1,6 +1,6 @@
 require 'logger'
 require 'rubber/environment'
-require 'rubber/instance'
+require 'rubber/configuration/cluster'
 require 'rubber/generator'
 
 module Rubber
@@ -53,9 +53,9 @@ module Rubber
         instance_storage = config['instance_storage']
         instance_storage_backup = config['instance_storage_backup']
         instance_storage ||= "file:#{@root}/instance-#{@env}.yml"
-        @instance = Instance.new(instance_storage, :backup => instance_storage_backup)
+        @instance = Cluster.new(instance_storage, :backup => instance_storage_backup)
       end
-      
+
       def environment
         @environment
       end
