+++ lib/rubber/configuration/role_item.rb
@@ -1,0 +1,71 @@
+module Rubber
+  module Configuration
+    # The configuration for a single role contained in the list
+    # of roles in InstanceItem
+    class RoleItem
+      attr_reader :name, :options
+
+      def self.expand_role_dependencies(roles, dependency_map, expanded=[])
+        roles = Array(roles)
+
+        if expanded.size == 0
+          common_deps = Array(dependency_map[RoleItem.new('common')])
+          roles.concat(common_deps)
+        end
+
+        roles.each do |role|
+          unless expanded.include?(role)
+            expanded << role
+            needed = dependency_map[role]
+            expand_role_dependencies(needed, dependency_map, expanded)
+          end
+        end
+
+        return expanded
+      end
+
+      def self.parse(str)
+        data = str.split(':')
+        role = Rubber::Configuration::RoleItem.new(data[0])
+        if data[1]
+          data[1].split(';').each do |pair|
+            p = pair.split('=')
+            val = case p[1]
+                    when 'true' then true
+                    when 'false' then false
+                    else p[1] end
+            role.options[p[0]] = val
+          end
+        end
+        return role
+      end
+
+      def to_s
+        str = @name
+        @options.each_with_index do |kv, i|
+          str += (i == 0 ? ':' : ';')
+          str += "#{kv[0]}=#{kv[1]}"
+        end
+        return str
+      end
+
+      def initialize(name, options={})
+        @name = name
+        @options = options || {}
+      end
+
+      def eql?(rhs)
+        rhs && @name == rhs.name && @options == rhs.options
+      end
+      alias == eql?
+
+      def hash
+        @name.hash
+      end
+
+      def <=>(rhs)
+        return @name <=> rhs.name
+      end
+    end
+  end
+end
